requirements.txt
aniso8601==3.0.2
asn1crypto==0.24.0
cffi==1.11.5
cryptography==1.9
Django==2.0.5
django-cors-headers==2.2.0
djangorestframework==3.8.2
freeze==1.0.10
graphene==2.1.3
graphene-django==2.2.0
graphql-core==2.1
graphql-relay==0.4.5
idna==2.6
promise==2.2.1
psycopg2==2.7.4
pycparser==2.18
PyJWT==1.6.4
pytz==2018.4
Rx==1.6.1
singledispatch==3.4.0.3
six==1.11.0
virtualenv==16.0.0
